# Optimization: use a more efficient regexp for hexadecimal check